redoing duckpan new files
