+++ lib/DDG/Spice/IndegoBikeshareLocations.pm
@@ -14,18 +14,18 @@ primary_example_queries 'indego philadelphia', 'bikeshare philadelphia';
 secondary_example_queries 'philly bikeshare';
 category 'geography';
 topics 'everyday', 'travel', 'entertainment', 'geography';
-code_url 'https://github.com/duckduckgo/zeroclickinfo-spice/blob/master/lib/DDG/Spice/Indego/IndegoBikeshareLocations.pm';
+code_url 'https://github.com/duckduckgo/zeroclickinfo-spice/blob/master/lib/DDG/Spice/PHLBikeshare/IndegoBikeshareLocations.pm';
 attribution github => 'AcriCAA',
             web  => ['http://www.coreyacri.com',  'Corey Acri'];
 
-triggers any => 'philly bikeshare', 'indego philadelphia', 'bikeshare', 'bike share';
+triggers any => 'philly bikeshare', 'indego philadelphia', 'bikeshare', 'bike share', 'go indego'. 'indego philly';
 
 spice to => 'https://api.phila.gov/bike-share-stations/v1';
 spice wrap_jsonp_callback => 1;
 spice proxy_cache_valid => '200 304 15m';
 
 handle remainder => sub {
-    $loc->city.' '.$_ =~ /(phl|philadelphia)/i;
+    $loc->city.' '.$_ =~ /(phl|philadelphia|philly)/i;
     return unless $1;
     return $1;
 };
