+++ share/spice/indego_bikeshare_locations/IndegoBikeshareLocations.handlebar
@@ -1,0 +1,4 @@
+<span>{{name}}</span>
+<a href="https://github.com/acricaa">
+   <img src="/iu/?u=https://avatars0.githubusercontent.com/u/4731092?v=3&s=460">
+</a>
