+++ share/spice/indego_bikeshare_locations/IndegoBikeshareLocations.pm
@@ -1,4 +1,4 @@
-package DDG::Spice::PHLBikeshare;
+package DDG::Spice::IndegoBikeshareLocations;
 
 
 use strict;
