--- share/spice/indego_bikeshare_locations/indego_bikeshare_locations.handlebars
@@ -1,4 +1,0 @@
-<span>{{name}}</span>
-<a href="http://some.url/profile/data1">
-   <img src="/iu/?u=Place a image url here">
-</a>
